Move payload live_os source under its directory
Payload sources will have the same structure as payloads.

There are two directories in the payload service: payloads and sources.
Each payload and source will have its own directory under one of these
two folders.
