+++ pyanaconda/modules/payload/factory.py
@@ -22,7 +22,7 @@ from pyanaconda.modules.payload.constants import HandlerType, SourceType
 from pyanaconda.modules.payload.payloads.dnf.dnf import DNFHandlerModule
 from pyanaconda.modules.payload.payloads.live_image.live_image import LiveImageHandlerModule
 from pyanaconda.modules.payload.payloads.live_os.live_os import LiveOSHandlerModule
-from pyanaconda.modules.payload.sources.live_os import LiveOSSourceModule
+from pyanaconda.modules.payload.sources.live_os.live_os import LiveOSSourceModule
 
 __all__ = ["HandlerFactory", "SourceFactory"]
