+++ pyanaconda/modules/payload/sources/live_os/initialization.py
