+++ pyanaconda/modules/payload/sources/live_os/live_os_interface.py
