+++ tests/nosetests/pyanaconda_tests/module_payload_live_os_test.py
@@ -33,7 +33,7 @@ from pyanaconda.modules.payload.base.initialization import UpdateBLSConfiguratio
 from pyanaconda.modules.payload.base.installation import InstallFromImageTask
 from pyanaconda.modules.payload.payloads.live_os.live_os import LiveOSHandlerModule
 from pyanaconda.modules.payload.payloads.live_os.live_os_interface import LiveOSHandlerInterface
-from pyanaconda.modules.payload.sources.live_os import LiveOSSourceModule
+from pyanaconda.modules.payload.sources.live_os.live_os import LiveOSSourceModule
 
 
 class LiveOSHandlerInterfaceTestCase(unittest.TestCase):
