+++ tests/nosetests/pyanaconda_tests/module_payload_test.py
@@ -41,7 +41,7 @@ from pyanaconda.modules.payload.payload import PayloadService
 from pyanaconda.modules.payload.payloads.dnf.dnf import DNFHandlerModule
 from pyanaconda.modules.payload.payloads.live_image.live_image import LiveImageHandlerModule
 from pyanaconda.modules.payload.payloads.live_os.live_os import LiveOSHandlerModule
-from pyanaconda.modules.payload.sources.live_os import LiveOSSourceModule
+from pyanaconda.modules.payload.sources.live_os.live_os import LiveOSSourceModule
 
 
 class PayloadInterfaceTestCase(TestCase):
